+++ CHANGELOG.md
@@ -18,6 +18,11 @@
   [Marcelo Fabri](https://github.com/marcelofabri)
   [#2249](https://github.com/realm/SwiftLint/issues/2249)
 
+* Add `anyobject_protocol` opt-in rule which suggests using `AnyObject` over
+  `class` for class-only protocols.  
+  [Ornithologist Coder](https://github.com/ornithocoder)
+  [#2283](https://github.com/realm/SwiftLint/issues/2283)
+
 #### Bug Fixes
 
 * Fix an issue with `control_statement` where commas in clauses prevented the